Colocando cada model em um app
