+++ sugestao/forms.py
@@ -2,7 +2,7 @@
 
 from django import forms
 from django.contrib.auth.models import User, Group
-from capacitaApp.models import *
+from sugestao.models import *
 
 class SugestaoForm(forms.ModelForm):
