+++ usuario/models.py
@@ -1,3 +1,4 @@
 from django.db import models
+from django.contrib.auth.models import User
 
 # Create your models here.
